html, css, js, jquery
a sample website for toy store.
problem statement goes like:
. Create a web application user interface with HTML, CSS, javaScript , bootstrap &
JQuery technology stack. This app should display different categories of toys for
children. At least 3 categories of toys should be there and each category consists of
a list of toys.
When the user clicks on a specific category that category list must be displayed.
When you click on a particular toy then it must display all the details of the toy - title,
price, description, category etc,.
Your app should consist of a home page, a Toys page, and a contact us page. Use
Visual Studio code IDE for the development of the application.